move configure_application_style into styles ui file
